mysite/requirements.txt
asn1crypto==0.24.0
astroid==2.2.4
autopep8==1.4.3
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.7
django-filter==2.0.0
django-rest-swagger==2.1.2
djangorestframework==3.9.2
idna==2.8
isort==4.3.12
itypes==1.1.0
Jinja2==2.10
jwt==0.6.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==5.4.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pymongo==3.7.2
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
